+++ packages/geo/src/lib/map/home-extent-button/home-extent-button.component.ts
@@ -36,11 +36,11 @@ export class HomeExtentButtonComponent {
       this.configService.getConfig('homeExtentButton');
 
     this.homeExtentButtonExtent =
-      this.extentOverride || homeExtentButtonOptions.homeExtButtonExtent;
+      this.extentOverride || homeExtentButtonOptions?.homeExtButtonExtent;
     this.homeExtentButtonCenter =
-      this.centerOverride || homeExtentButtonOptions.homeExtButtonCenter;
+      this.centerOverride || homeExtentButtonOptions?.homeExtButtonCenter;
     this.homeExtentButtonZoom =
-      this.zoomOverride || homeExtentButtonOptions.homeExtButtonZoom;
+      this.zoomOverride || homeExtentButtonOptions?.homeExtButtonZoom;
 
     // priority over extent if these 2 properties are defined;
     if (this.centerOverride && this.zoomOverride) {